requirements.txt
Django==4.1.1
psycopg2==2.9.3
gunicorn==19.9.0
asgiref==3.5.2
django-restframework==0.0.1
sqlparse==0.4.3
tzdata==2022.4
django-recaptcha==3.0.0